Added all missing 2020 packages

+++ scripts/vfx/build_alembic.sh
@@ -5,9 +5,9 @@ set -ex
 mkdir alembic
 cd alembic
 
-#if [ ! -f $DOWNLOADS_DIR/hdf5-${HDF5_VERSION}.tar.gz ]; then
+if [ ! -f $DOWNLOADS_DIR/hdf5-${HDF5_VERSION}.tar.gz ]; then
     curl --location https://support.hdfgroup.org/ftp/HDF5/releases/hdf5-1.8/hdf5-${HDF5_VERSION}/src/hdf5-${HDF5_VERSION}.tar.gz -o $DOWNLOADS_DIR/hdf5-${HDF5_VERSION}.tar.gz
-#fi
+fi
 
 tar -zxf $DOWNLOADS_DIR/hdf5-${HDF5_VERSION}.tar.gz
 cd hdf5-${HDF5_VERSION}
@@ -26,13 +26,22 @@ if [ ! -f $DOWNLOADS_DIR/alembic-${ALEMBIC_VERSION}.tar.gz ]; then
 fi
 tar -zxf $DOWNLOADS_DIR/alembic-${ALEMBIC_VERSION}.tar.gz
 cd alembic-${ALEMBIC_VERSION}
+
+# Boost Python3 not found by cmake for PyAlembic as of Alembic 1.7.12
+if [[ $PYTHON_VERSION == 2.7* ]]; then
+    USE_PYALEMBIC=TRUE
+else
+    USE_PYALEMBIC=FALSE
+fi
+
 cmake \
     -D CMAKE_INSTALL_PREFIX=${ASWF_INSTALL_PREFIX} \
     -D CMAKE_PREFIX_PATH=${ASWF_INSTALL_PREFIX} \
     -D BOOST_ROOT=${ASWF_INSTALL_PREFIX} \
     -D ILMBASE_ROOT=${ASWF_INSTALL_PREFIX} \
+    -D Python_ADDITIONAL_VERSIONS=3 \
     -D USE_PYILMBASE=TRUE \
-    -D USE_PYALEMBIC=TRUE \
+    -D USE_PYALEMBIC=${USE_PYALEMBIC} \
     -D USE_ARNOLD=FALSE \
     -D USE_PRMAN=FALSE \
     -D USE_MAYA=FALSE \
